refactor(eventTrigger): 优化事件触发保存逻辑
- 新增 NotifyProperties 中最大重试次数配置，默认值设为5
- 在 EventPushServiceImpl 中增加 saveByEventTrigger 方法的初步实现，记录事件推送相关信息
- EventTriggerServiceImpl 新增 save 方法，结合事件类型校验，完善事件触发保存及事件推送逻辑

+++ mdp-platform/md-open/open-pojo/src/main/java/top/mddata/open/admin/vo/EventTriggerVo.java
@@ -47,7 +47,7 @@ public class EventTriggerVo implements Serializable {
      * 事件ID
      */
     @Schema(description = "事件ID")
-    private Integer eventId;
+    private Long eventId;
 
     /**
      * 触发时间
